requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.12
django-allauth==0.42.0
django-cors-headers==3.3.0
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
idna==2.9
Markdown==3.2.2
oauthlib==3.1.0
psycopg2==2.8.5
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9